refactor(ptt): extract PttCard sub-component
